Suppression RTC, remplacé par gb.getTimePerFrame

+++ Platformer/HighScore.cpp
@@ -1,9 +1,10 @@
 // author: chris-scientist
 // created at: 08/02/2019
+// updated at: 09/02/2019
 
 #include "HighScore.h"
 
-void initHighScore(HighScore &aHighScore, char * aName, uint8_t aScore) {
+void initHighScore(HighScore &aHighScore, char * aName, int32_t aScore) {
   strncpy(aHighScore.nameOfScore, aName, 5);
   aHighScore.score = aScore;
 }
@@ -12,7 +13,7 @@ void initHighScore(HighScore &aHighScore, char * aName, uint8_t aScore) {
 // Si a == b alors retourner 0
 // Si a > b alors retourner 1
 // Si a < b alors retourner -1
-int8_t compareTime(const HighScore &aScore, const uint8_t aTimeInSeconds) {
+int8_t compareTime(const HighScore &aScore, const int32_t aTimeInSeconds) {
   if(aScore.score == aTimeInSeconds) {
     return 0;
   } else if(aScore.score > aTimeInSeconds) {
